requirements.txt
streamlit==1.29.0
jax==0.4.13
jaxlib==0.4.13
flax==0.7.0
tensorflow==2.13.0
tensorflow-datasets==4.9.2
optax==0.1.7
Pillow==10.1.0
numpy==1.24.3
scipy==1.11.3
joblib==1.3.2